+++ src/Page/ContentEdit.php
@@ -102,6 +102,11 @@ class ContentEdit
 			'pageID'    => $page->id,
 		]);
 
+		$this->_dispatcher->dispatch(
+			Event\ContentEvent::EDIT,
+			new Event\ContentEvent($page, $content)
+		);
+
 		return $this->_transaction->commit();
 	}
 
@@ -149,7 +154,13 @@ class ContentEdit
 		foreach ($content as $part) {
 			if ($part instanceof Field\RepeatableContainer) {
 				foreach ($part as $i => $group) {
-					$this->_appendGroup($group, $i);
+					$seq = $i;
+
+					if (isset($group->{Field\ContentBuilder::SEQUENCE_FIELD})) {
+						$seq = $group->{Field\ContentBuilder::SEQUENCE_FIELD}->getValue();
+					}
+
+					$this->_appendGroup($group, (int) $seq);
 				}
 			}
 			elseif ($part instanceof Field\Group) {